# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/categories": {
      "get": {
        "description": "Returns categories with optional status + search filters and pagination. Admin-only.",
        "operationId": "AdminCategoriesController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status.",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "DEACTIVATED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive substring match on name.",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number.",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Items per page (1-100).",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of categories."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "List categories (admin)",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminCategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "409": {
            "description": "A category with this name already exists."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Create a new category (admin)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/categories/{id}": {
      "get": {
        "operationId": "AdminCategoriesController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category id (cuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category found."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "Category not found."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Get a category by id (admin)",
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "AdminCategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category id (cuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated."
          },
          "400": {
            "description": "No fields to update."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "Category not found."
          },
          "409": {
            "description": "Name collision with another ACTIVE category."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Update a category (partial)",
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminCategoriesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category id (cuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Category deleted."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "Category not found."
          },
          "409": {
            "description": "Category is in use by one or more doctors."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Hard-delete a category",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/categories/{id}/deactivate": {
      "patch": {
        "operationId": "AdminCategoriesController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category id (cuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category deactivated (idempotent)."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "Category not found."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Soft-deactivate a category",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vezeeta Backend API",
    "description": "REST API for the Vezeeta backend.\n\n## Auth\n\nAuth is handled by [Better Auth](https://www.better-auth.com/) and exposed under `/api/auth/*`.\nAfter `sign-in/email` or `sign-in/phone-number`, the server sets an HTTP-only session cookie.\nThe frontend must send the cookie on every subsequent request (use `credentials: \"include\"` in fetch / `withCredentials: true` in axios).\n\nFor the social OAuth flow, the browser hits `GET /api/auth/oauth/start?provider=google` (or `facebook`) — a Better Auth standard route — and follows the 302 redirect.\n\nUse the **Authorize** button below to paste the session cookie value if you want to call protected endpoints from Swagger UI.\n\n## Tags\n\n- `auth` — current-session, social linking, health\n- `doctors` — public doctor catalog\n- `categories` — public category dropdown\n- `slots` — public slot picker for a doctor\n- `doctor-services` — per-doctor service catalog (admin)\n- `appointments` — patient booking flow\n- `reviews` — patient reviews + public doctor rating\n- `medical-records` — patient-facing read of their records\n- `notifications` — in-app inbox\n- `admin` — admin-only management (requires the `admin` role)",
    "version": "0.0.1",
    "contact": {},
    "x-generated-at": "2026-07-22T17:35:45.274Z"
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "vezeta.session_token",
        "description": "Better Auth session cookie set by /api/auth/* (prefix configured in src/auth/auth.ts)."
      }
    },
    "schemas": {
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Category display name. Trimmed of leading/trailing whitespace.",
            "minLength": 1,
            "maxLength": 100,
            "example": "Cardiology"
          },
          "status": {
            "type": "string",
            "description": "Category status. Defaults to ACTIVE when omitted.",
            "enum": [
              "ACTIVE",
              "DEACTIVATED"
            ]
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "New category name. Trimmed of leading/trailing whitespace.",
            "minLength": 1,
            "maxLength": 100
          },
          "status": {
            "type": "string",
            "description": "New category status.",
            "enum": [
              "ACTIVE",
              "DEACTIVATED"
            ]
          }
        }
      },
...
    }
  }
}